Started turning the cluster into modules.
Much better reusability, and most of its functionality was related to other responsibilities anyway.
